+++ apps/web/src/app/[locale]/guild/@modal/default.tsx
@@ -1,0 +1,3 @@
+export default function GuildModal() {
+  return null;
+}
