+++ apps/web/src/app/[locale]/guild/detail/[id]/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { useRouter } from '@/i18n/routing';
+
+import { GuildDetail } from '../../GuildDetail';
+
+export default function GuildDetailPage({ params }: { params: { id: string } }) {
+  const router = useRouter();
+
+  return <GuildDetail guildId={params.id} onJoin={() => router.push(`/guild/detail/${params.id}/join`)} />;
+}
