+++ apps/web/src/app/[locale]/guild/layout.tsx
@@ -2,11 +2,12 @@ import { css } from '_panda/css';
 
 import GNB from '@/components/GNB/GNB';
 
-export default function GuildLayout({ children }: { children: React.ReactNode }) {
+export default function GuildLayout({ children, modal }: { children: React.ReactNode; modal: React.ReactNode }) {
   return (
     <>
       <GNB />
       <div className={containerStyle}>{children}</div>
+      {modal}
     </>
   );
 }
@@ -15,7 +16,7 @@ const containerStyle = css({
   width: '100%',
   background: 'linear-gradient(180deg, #000 0%, #004875 38.51%, #005B93 52.46%, #006FB3 73.8%, #0187DB 100%)',
   minHeight: 'calc(100vh - 60px)',
-  height: 'calc(100vh - 60px)',
+  height: 'fit-content',
   overflow: 'hidden',
   position: 'relative',
